feat(movieDetailsModule) 🚧 Convert the movie details module into a separated local package.
