+++ MoviesList-MVVM-SwiftUI/Modules/MovieDetailsModule/Sources/MovieDetailsModule/Data/Repository/MovieDetailsRepository.swift
@@ -6,6 +6,7 @@
 //
 
 import Combine
+import MANetwork
 
 final
 class MovieDetailsRepository {
