+++ MoviesList-MVVM-SwiftUI/Modules/MovieDetailsModule/Sources/MovieDetailsModule/Data/Repository/MovieDetailsRepositoryProtocol.swift
@@ -6,6 +6,7 @@
 //
 
 import Combine
+import MANetwork
 
 protocol MovieDetailsRepositoryProtocol: MovieDetailsRepositoryGettable { }
