+++ MoviesList-MVVM-SwiftUI/Modules/MovieDetailsModule/Sources/MovieDetailsModule/Presentation/ViewModel/MovieDetailsViewModel.swift
@@ -7,6 +7,7 @@
 
 import SwiftUI
 import Combine
+import Commons
 
 final class MovieDetailsViewModel: ObservableObject {
   // MARK: - Vars
