+++ MoviesList-MVVM-SwiftUI/Modules/MovieDetailsModule/Tests/MovieDetailsModuleTests/MovieDetailsModuleTests.swift
@@ -1,0 +1,12 @@
+import XCTest
+@testable import MovieDetailsModule
+
+final class MovieDetailsModuleTests: XCTestCase {
+    func testExample() throws {
+        // XCTest Documentation
+        // https://developer.apple.com/documentation/xctest
+
+        // Defining Test Cases and Test Methods
+        // https://developer.apple.com/documentation/xctest/defining_test_cases_and_test_methods
+    }
+}
